# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  phone: string;
  name: string;
  createdAt: string;
}

export interface PublicUser {
  name: string;
  phone: string;
}

export interface Wallet {
  balanceCents: number;
}

export interface SessionInfo {
  user: User;
  wallet: Wallet;
}

export interface OtpRequestInput {
  /**
     * @minLength 6
     * @maxLength 15
     */
  phone: string;
}

export interface OtpRequestResult {
  phone: string;
  /** Simulated OTP, shown directly since no real SMS gateway is connected in this demo. */
  otp: string;
  expiresInSeconds: number;
}

export interface OtpVerifyInput {
  phone: string;
  otp: string;
  /**
     * Optional display name to set on first login.
     * @nullable
     */
  name?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  add_money: 'add_money',
  transfer_out: 'transfer_out',
  transfer_in: 'transfer_in',
  recharge: 'recharge',
  bill_payment: 'bill_payment',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amountCents: number;
  status: string;
  /** @nullable */
  counterpartyName?: string | null;
  /** @nullable */
  counterpartyPhone?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface WalletMutationResult {
  wallet: Wallet;
  transaction: Transaction;
}

export type AddMoneyInputMethod = typeof AddMoneyInputMethod[keyof typeof AddMoneyInputMethod];


export const AddMoneyInputMethod = {
  upi: 'upi',
  debit_card: 'debit_card',
  credit_card: 'credit_card',
  netbanking: 'netbanking',
} as const;

export interface AddMoneyInput {
  /** @minimum 100 */
  amountCents: number;
  method: AddMoneyInputMethod;
}

export interface TransferInput {
  toPhone: string;
  /** @minimum 100 */
  amountCents: number;
  /** @nullable */
  note?: string | null;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  nickname?: string | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 6 */
  phone: string;
  /** @nullable */
  nickname?: string | null;
}

export type BillerCategory = typeof BillerCategory[keyof typeof BillerCategory];


export const BillerCategory = {
  mobile_recharge: 'mobile_recharge',
  dth: 'dth',
  electricity: 'electricity',
  water: 'water',
  gas: 'gas',
  broadband: 'broadband',
  insurance: 'insurance',
  loan_payment: 'loan_payment',
} as const;

export interface Biller {
  id: string;
  category: BillerCategory;
  name: string;
}

export interface RechargeInput {
  billerId: string;
  /** @minLength 1 */
  accountNumber: string;
  /** @minimum 100 */
  amountCents: number;
}

export interface CategorySpend {
  category: string;
  totalCents: number;
}

export interface DashboardSummary {
  balanceCents: number;
  totalAddedCents: number;
  totalSpentCents: number;
  recentTransactions: Transaction[];
  spendByCategory: CategorySpend[];
}

export type ListTransactionsParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
};

export type LookupUserParams = {
phone: string;
};

export type ListBillersParams = {
category?: BillerCategory;
};