# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request a one-time login code for a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequestInput"
      responses:
        "200":
          description: OTP generated (simulated delivery)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify the one-time code and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentSession
      tags: [auth]
      summary: Get the current logged-in user and wallet
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /billers:
    get:
      operationId: listBillers
      tags: [billers]
      summary: List available recharge and bill payment providers
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BillerCategory"
      responses:
        "200":
          description: Billers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Biller"
  /recharge:
    post:
      operationId: createRecharge
      tags: [billers]
      summary: Pay a recharge or bill using a biller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "200":
          description: Payment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletMutationResult"
        "400":
          description: Invalid recharge (unknown biller, insufficient balance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, phone, name, createdAt]
      properties:
        id: { type: integer }
        phone: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
...
    Wallet:
      type: object
      required: [balanceCents]
      properties:
        balanceCents: { type: integer }
    SessionInfo:
      type: object
      required: [user, wallet]
      properties:
        user:
          $ref: "#/components/schemas/User"
        wallet:
          $ref: "#/components/schemas/Wallet"
    OtpRequestInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 6
          maxLength: 15
    OtpRequestResult:
      type: object
      required: [phone, otp, expiresInSeconds]
      properties:
        phone: { type: string }
        otp:
          type: string
          description: Simulated OTP, shown directly since no real SMS gateway is connected in this demo.
        expiresInSeconds: { type: integer }
    OtpVerifyInput:
      type: object
      required: [phone, otp]
      properties:
        phone: { type: string }
        otp: { type: string }
        name:
          type: ["string", "null"]
          description: Optional display name to set on first login.
    TransactionType:
      type: string
      enum: [add_money, transfer_out, transfer_in, recharge, bill_payment]
    Transaction:
      type: object
      required: [id, type, amountCents, status, createdAt]
      properties:
        id: { type: integer }
        type:
          $ref: "#/components/schemas/TransactionType"
        amountCents: { type: integer }
        status: { type: string }
        counterpartyName:
          type: ["string", "null"]
        counterpartyPhone:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
    WalletMutationResult:
      type: object
      required: [wallet, transaction]
      properties:
        wallet:
          $ref: "#/components/schemas/Wallet"
        transaction:
          $ref: "#/components/schemas/Transaction"
...
    BillerCategory:
      type: string
      enum:
        [
          mobile_recharge,
          dth,
          electricity,
          water,
          gas,
          broadband,
          insurance,
          loan_payment,
        ]
    Biller:
      type: object
      required: [id, category, name]
      properties:
        id: { type: string }
        category:
          $ref: "#/components/schemas/BillerCategory"
        name: { type: string }
    RechargeInput:
      type: object
      required: [billerId, accountNumber, amountCents]
      properties:
        billerId: { type: string }
        accountNumber:
          type: string
          minLength: 1
        amountCents:
          type: integer
          minimum: 100
...