# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';
import type { Wallet } from './wallet';

export interface SessionInfo {
  user: User;
  wallet: Wallet;
}